+++ .gitignore
@@ -17,3 +17,6 @@ node_modules
 # Mac
 .DS_Store
 /.next/
+
+# Sensitive config
+.env.development.local
