+++ package.json
@@ -18,6 +18,7 @@
     "@emotion/react": "11.13.3",
     "@emotion/styled": "11.13.0",
     "@fontsource/inter": "5.1.0",
+    "@mapbox/search-js-react": "^1.0.0-beta.22",
     "@types/node": "20.2.4",
     "@types/react": "18.2.7",
     "@types/react-dom": "18.2.4",
